Add additional download route tests
